use dotenv babel-register polyfill
