+++ package.json
@@ -8,7 +8,7 @@
     "lint": "eslint client server",
     "build": "NODE_ENV=production webpack --config webpack.config.prod.js --progress",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "fetch-examples": "BABEL_DISABLE_CACHE=1 NODE_ENV=development nodemon fetch-examples.js"
+    "fetch-examples": "node fetch-examples.js"
   },
   "main": "index.js",
   "author": "Cassie Tarakajian",
