+++ server/models/user.js
@@ -1,5 +1,4 @@
-// import mongoose from 'mongoose';
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 const Schema = mongoose.Schema;
 const bcrypt = require('bcrypt-nodejs');
 
@@ -59,5 +58,4 @@ userSchema.methods.comparePassword = function comparePassword(candidatePassword,
   });
 };
 
-// export default mongoose.model('User', userSchema);
-module.exports = mongoose.model('User', userSchema);
+export default mongoose.model('User', userSchema);
